change all of the Ints to Bigints to comply with BotAPI 5.4
Yes, I know, it's to late but anyways

+++ tg_bot/modules/sql/api_sql.py
@@ -17,13 +17,13 @@
 
 
 
-from sqlalchemy import Column, UnicodeText, Integer
+from sqlalchemy import Column, UnicodeText, Integer, BigInteger
 from tg_bot.modules.sql import SESSION, BASE
 
 
 class Api(BASE):
     __tablename__ = "api"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     key = Column(UnicodeText)
 
     def __init__(self, user_id, key=None):
